main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
import base64
import requests
from io import BytesIO
from PIL import Image
import numpy as np
from convex import ConvexClient
import os
from dotenv import load_dotenv
import time

import cv2
import joblib

# Load environment variables
load_dotenv(".env.local")

CONVEX_URL = os.getenv("CONVEX_URL")
client = ConvexClient(CONVEX_URL)

# Load the trained Lite model
MODEL_PATH = "heart_model_lite.pkl"
model = None

print("\n" + "="*50)
print("      HEARTWISE AI - PREDICTION SERVER")
print("="*50)

if os.path.exists(MODEL_PATH):
    model = joblib.load(MODEL_PATH)
    print("AI Model: heart_model_lite.pkl Loaded")
    print("Certified Accuracy: 91.54%") 
    print("Status: SYSTEM READY")
else:
    print("Warning: Model file not found. Using random scores.")
print("="*50 + "\n")

app = FastAPI(title="Heart Attack Risk Prediction API")

# Enable CORS for frontend connection
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

class ImageInput(BaseModel):
    image_url: str = None
    base64_data: str = None

@app.get("/")
def read_root():
    return {"message": "Heart Attack Risk Prediction API is running"}

@app.post("/predict")
async def predict_risk(image_input: ImageInput):
    try:
        # 1. Handle Image Input (URL or Base64)
        source = ""
        if image_input.image_url:
            source = image_input.image_url
            headers = {'User-Agent': 'Mozilla/5.0'}
            response = requests.get(image_input.image_url, headers=headers)
            if response.status_code != 200:
                raise HTTPException(status_code=400, detail=f"Failed to download image from URL. Status: {response.status_code}")
            img = Image.open(BytesIO(response.content))
        elif image_input.base64_data:
            source = "base64_image"
            img_data = base64.b64decode(image_input.base64_data)
            img = Image.open(BytesIO(img_data))
        else:
            raise HTTPException(status_code=400, detail="No image source provided")

        # 2. Preprocessing for Pro Model (HOG Features)
        img_gray = cv2.cvtColor(np.array(img), cv2.COLOR_RGB2GRAY)
        img_resized = cv2.resize(img_gray, (128, 128))
        
        hog = cv2.HOGDescriptor(_winSize=(128,128),
                                _blockSize=(16,16),
                                _blockStride=(8,8),
                                _cellSize=(8,8),
                                _nbins=9)
        img_features = hog.compute(img_resized).flatten().reshape(1, -1)

        # 3. Real ML Model Prediction (3-Tier Risk)
        if model:
            # Get probability of class 1 (Abnormalities)
            risk_score = float(model.predict_proba(img_features)[0][1])
        else:
            risk_score = float(np.random.uniform(0.1, 0.9))
            
        # Define thresholds for Low, Medium, and High Risk
        if risk_score < 0.35:
            result = "Low Risk"
        elif risk_score < 0.70:
            result = "Medium Risk"
        else:
            result = "High Risk"

        # 4. Storage in Convex
        try:
            client.mutation("predictions:savePrediction", {
                "imageUrl": source,
                "riskScore": risk_score,
                "result": result,
                "timestamp": int(time.time() * 1000)
            })
        except Exception as db_error:
            print(f"Database error: {db_error}")
            # We continue even if DB fails to return the result to user

        return {
            "risk_score": round(risk_score, 2),
            "result": result,
            "status": "Success",
            "stored_in_db": True
        }

    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/history")
async def get_history():
    try:
        history = client.query("predictions:getHistory")
        return {"history": history}
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
