requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
requests
python-multipart
python-dotenv
convex
